refactor(ocale): migrate places publish wire code into domains/place-publishing
Moves the legacy /universes/v1/.../versions publish/save Operation wire
code (builder, parser, wire types, rate-limit constant, parsed shapes,
and signature validation helpers) for the Places Resource into a new
per-shape sub-tree under src/domains/place-publishing/places/, following
ADR-022. The Domain folder name follows the drop-legacy-and-version
rule, departing from the literal universes prefix because that would
collide with domains/cloud-v2/universes/; place-publishing reflects the
sole purpose of the URL prefix and matches Roblox's docs guide name.

resources/places/ now contains only client.ts, client.example.spec.ts,
and index.ts. The client binds publish/save and update from their
respective Domain sub-trees. Public surface, wire requests, parsed
shapes, and operation keys (places.publishVersion, places.update) are
preserved end-to-end; conformance and integration tests pass without
modification.
