+++ packages/open-cloud/src/domains/place-publishing/places/operations.ts
@@ -1,4 +1,4 @@
-import type { OperationLimit } from "../../internal/http/rate-limit-queue.ts";
+import type { OperationLimit } from "../../../internal/http/rate-limit-queue.ts";
 
 /**
  * Per-second request ceiling for publishing or saving a place version,
