+++ packages/open-cloud/src/domains/place-publishing/places/parsers.spec.ts
@@ -1,6 +1,6 @@
 import { assert, describe, expect, it } from "vitest";
 
-import { ApiError } from "../../errors/api-error.ts";
+import { ApiError } from "../../../errors/api-error.ts";
 import { parsePublishResponse } from "./parsers.ts";
 
 describe(parsePublishResponse, () => {
