+++ packages/open-cloud/src/domains/place-publishing/places/parsers.ts
@@ -1,7 +1,7 @@
-import type { HttpResponse } from "../../client/types.ts";
-import { ApiError } from "../../errors/api-error.ts";
-import { isRecord } from "../../internal/utils/is-record.ts";
-import type { Result } from "../../types.ts";
+import type { HttpResponse } from "../../../client/types.ts";
+import { ApiError } from "../../../errors/api-error.ts";
+import { isRecord } from "../../../internal/utils/is-record.ts";
+import type { Result } from "../../../types.ts";
 import type { PlaceVersion } from "./types.ts";
 import type { PlaceVersionWire } from "./wire.ts";
