+++ packages/open-cloud/src/domains/place-publishing/places/signatures.ts
