+++ packages/open-cloud/src/resources/places/index.ts
@@ -1,3 +1,6 @@
 export type { Place, UpdatePlaceParameters } from "../../domains/cloud-v2/places/types.ts";
+export type {
+	PlaceVersion,
+	PublishParameters,
+} from "../../domains/place-publishing/places/types.ts";
 export { PlacesClient } from "./client.ts";
-export type { PlaceVersion, PublishParameters } from "./types.ts";
